requirements.txt
robotframework==6.1.1
robotframework-seleniumlibrary==6.1.1
robotframework-datadriver==1.9.0
robotframework-pabot==2.16.0
robotframework-requests==0.9.5
